requirements.txt
pypdf
langchain-text-splitters
ollama
pinecone[grpc]
python-dotenv
